backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Александр Меньшиков <menshikov.a.v@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
python-decouple = "^3.8"
psycopg2 = "^2.9.10"
djangorestframework = "^3.15.2"
gunicorn = "^23.0.0"
gevent = "^24.10.3"
drf-spectacular = "^0.27.2"
django-cors-headers = "^4.5.0"
huey = "^2.5.2"
redis = "^5.2.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
model-bakery = "^1.20.0"

[tool.black]
line-length = 120
target-version = ['py311']
skip-string-normalization = true
extend-exclude = '''
(
  migrations
)
'''

[tool.isort]
# Global
multi_line_output = 3
include_trailing_comma = true
line_length = 120
lines_after_imports = 2
lines_between_sections = 1
lines_between_types = 0
force_alphabetical_sort_within_sections = true
ensure_newline_before_comments = true
combine_as_imports = true
combine_star = true
skip_gitignore = true
balanced_wrapping = true
force_grid_wrap = 2
# Local
src_paths = [
    "src",
]
skip_glob = [
    ".git",
    ".idea",
    ".vscode",
    "venv",
]
# Known
known_django = "django"
known_rest_framework = "rest_framework"
# Sections
sections = [
    "FUTURE",
    "STDLIB",
    "DJANGO",
    "REST_FRAMEWORK",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
